move global reducer to own folder
